backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "herbo-nutra-backend"
version = "0.1.0"
description = "FastAPI backend for Herbo Nutra Extract website"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "alembic>=1.16.0",
  "email-validator>=2.2.0",
  "fastapi>=0.131.0",
  "httpx>=0.28.0",
  "pydantic-settings>=2.11.0",
  "PyJWT>=2.10.0",
  "python-multipart>=0.0.20",
  "SQLAlchemy>=2.0.40",
  "psycopg[binary]>=3.2.0",
  "passlib>=1.7.4",
  "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
  "ignore::DeprecationWarning",
]
